// projects/moba-rad/client/src/netcode/runtimeDiagnostics.ts
import type { AckDiagnosticsSnapshot } from './ackDiagnostics';

export interface NetcodeDiagnosticsSnapshot extends AckDiagnosticsSnapshot {
  authoritySynced: boolean;
  authorityStateInFlight: boolean;
  authorityMayBeMoving: boolean;
  localSimulationActive: boolean;
  localTick: number;
  serverTickEstimate: number;
  predictionLeadTicks: number;
  roundTripMs: number;
  jitterMs: number;
  lastServerSeq: number;
  statePacketsReceived: number;
  acceptedStatePackets: number;
  staleStatePackets: number;
  rejectedStatePackets: number;
  droppedStatePackets: number;
  correctionCount: number;
  maxCorrectionDistance: number;
  smoothedCorrectionCount: number;
  reconciliationRatePerSecond: number;
  inputPacketsSent: number;
  inputResendPackets: number;
  transportFailures: number;
  lastTransportError: string;
  authorityTimeouts: number;
  authorityStateRequests: number;
  peerCount: number;
  peerRecordCount: number;
  maxPeers: number;
  inputQueueSlots: number;
  pendingMoveInputs: number;
  pendingCastInputs: number;
  peerConnected: boolean;
  lateInputs: number;
  futureInputs: number;
  duplicateInputs: number;
  overwrittenInputs: number;
  lastAuthorityClientSeq: number;
  lastAuthorityAppliedSeq: number;
  lastAuthorityAppliedAckBits: number;
  lastAuthorityStatus: string;
  lastCorrectionReason: string;
  avatarRecordCount: number;
  remoteAvatarCount: number;
  projectileRecordCount: number;
  projectileImpactRecordCount: number;
  remoteAvatarPoolActive: number;
  remoteAvatarPoolIdle: number;
  projectilePoolActive: number;
  projectilePoolIdle: number;
  impactPoolActive: number;
  impactPoolIdle: number;
}

export function createNetcodeDiagnosticsSnapshot(): NetcodeDiagnosticsSnapshot {
  return {
    highestAck: 0,
    inspectedPackets: 0,
    missingPackets: 0,
    lossRatio: 0,
    recommendedInputDelayTicks: 0,
    authoritySynced: false,
    authorityStateInFlight: false,
    authorityMayBeMoving: false,
    localSimulationActive: false,
    localTick: 0,
    serverTickEstimate: 0,
    predictionLeadTicks: 0,
    roundTripMs: 0,
    jitterMs: 0,
    lastServerSeq: 0,
    statePacketsReceived: 0,
    acceptedStatePackets: 0,
    staleStatePackets: 0,
    rejectedStatePackets: 0,
    droppedStatePackets: 0,
    correctionCount: 0,
    maxCorrectionDistance: 0,
    smoothedCorrectionCount: 0,
    reconciliationRatePerSecond: 0,
    inputPacketsSent: 0,
    inputResendPackets: 0,
    transportFailures: 0,
    lastTransportError: 'none',
    authorityTimeouts: 0,
    authorityStateRequests: 0,
    peerCount: 0,
    peerRecordCount: 0,
    maxPeers: 0,
    inputQueueSlots: 0,
    pendingMoveInputs: 0,
    pendingCastInputs: 0,
    peerConnected: false,
    lateInputs: 0,
    futureInputs: 0,
    duplicateInputs: 0,
    overwrittenInputs: 0,
    lastAuthorityClientSeq: 0,
    lastAuthorityAppliedSeq: 0,
    lastAuthorityAppliedAckBits: 0,
    lastAuthorityStatus: 'none',
    lastCorrectionReason: 'none',
    avatarRecordCount: 0,
    remoteAvatarCount: 0,
    projectileRecordCount: 0,
    projectileImpactRecordCount: 0,
    remoteAvatarPoolActive: 0,
    remoteAvatarPoolIdle: 0,
    projectilePoolActive: 0,
    projectilePoolIdle: 0,
    impactPoolActive: 0,
    impactPoolIdle: 0,
  };
}